Add helpers for resizing
This adds functions to set the viewport to a surface or a window, and
also adds a helper for calculating projections.
