+++ Cargo.toml
@@ -30,7 +30,7 @@ maintenance = { status = "actively-developed" }
 [dependencies]
 bytemuck = "1.0"
 gestalt = { version = "0.1", optional = true }
-golem = "0.1.0-alpha6"
+golem = "0.1.0"
 image = { version = "0.22", default-features = false, features = ["png_codec", "jpeg"] }
 log = "0.4"
 mint = "0.5.3"
